extract IntentEncoder to reuse
